requirements.txt
dimod==0.12.18
dwave-cloud-client==0.13.2
dwave-neal==0.5.9
dwave-networkx==0.8.12
dwave-optimization==0.4.3
dwave-preprocessing==0.6.7
dwave-samplers==1.2.0
dwave-system==1.28.0
matplotlib==3.3.4
matplotlib-inline==0.1.3
seaborn==0.11.2
networkx==2.8.8
opt_einsum==3.4.0
packaging==25.0
tornado==6.5.2
tqdm==4.64.0
metis==0.2a5
PyMetis==2023.1.1
minorminer==0.2.16
numpy==1.22.4
scipy==1.10.1
numba==0.55.2
scikit-learn==1.1.3
torch==2.5.1
torchvision==0.20.1
